feat: change files, because have a lot of bags with modules & fix errors
