+++ .env
@@ -1,2 +1,2 @@
-PORT=3000
+PORT=5000
 NUM_WORKERS=4
